everything arround analyzing yaml files
