+++ tools/serverpod_cli/lib/analyzer.dart
@@ -2,11 +2,11 @@
 library analyzer;
 
 export 'src/generator/config.dart' show GeneratorConfig;
-export 'src/analyzer/file_analyzer.dart' show ProtocolFileAnalyzer;
+export 'src/analyzer/yaml/file_analyzer.dart' show ProtocolYamlFileAnalyzer;
 export 'src/generator/code_analysis_collector.dart' show CodeAnalysisCollector;
 export 'package:source_span/source_span.dart'
     show SourceSpanException; // The collector uses them.
-export 'src/generator/protocol_definition.dart'
+export 'src/analyzer/protocol_definition.dart'
     show
         ProtocolFileDefinition,
         ClassDefinition,
