+++ tools/serverpod_cli/lib/database.dart
@@ -1,0 +1,13 @@
+
+library database;
+
+export 'src/analyzer/protocol_definition.dart'
+    show
+        ProtocolFileDefinition,
+        ClassDefinition,
+        FieldDefinition,
+        FieldScope,
+        IndexDefinition,
+        EnumDefinition;
+export 'src/generator/types.dart' show TypeDefinition;
+
