+++ tools/serverpod_cli/lib/src/generator/class_generator_dart.dart
@@ -1,9 +1,10 @@
 import 'package:built_collection/built_collection.dart';
 import 'package:code_builder/code_builder.dart';
+import 'package:serverpod_cli/src/analyzer/yaml/definitions.dart';
 
 import 'class_generator.dart';
 import 'config.dart';
-import 'protocol_definition.dart';
+import '../analyzer/protocol_definition.dart';
 
 String serverpodUrl(bool serverCode) {
   return serverCode
