+++ tools/serverpod_cli/lib/src/generator/types.dart
@@ -1,13 +1,14 @@
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:code_builder/code_builder.dart';
+import 'package:serverpod_cli/src/analyzer/yaml/definitions.dart';
 import 'package:serverpod_shared/serverpod_shared.dart';
 import 'package:source_span/source_span.dart';
 import 'package:path/path.dart' as p;
 
 import 'class_generator_dart.dart';
 import 'config.dart';
-import 'protocol_definition.dart';
+import '../analyzer/protocol_definition.dart';
 
 /// Contains information about the type of fields, arguments and return values.
 class TypeDefinition {
